+++ modules/etc/image_compress_etc.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  image_etc.h                                                          */
+/*  image_compress_etc.h                                                 */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,9 +28,9 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifndef IMAGE_ETC1_H
-#define IMAGE_ETC1_H
+#ifndef IMAGE_COMPRESS_ETC_H
+#define IMAGE_COMPRESS_ETC_H
 
 void _register_etc_compress_func();
 
-#endif // IMAGE_ETC_H
+#endif // IMAGE_COMPRESS_ETC_H
