+++ modules/pvr/register_types.cpp
@@ -30,6 +30,7 @@
 
 #include "register_types.h"
 
+#include "image_compress_pvrtc.h"
 #include "texture_loader_pvr.h"
 
 static Ref<ResourceFormatPVR> resource_loader_pvr;
@@ -37,6 +38,8 @@ static Ref<ResourceFormatPVR> resource_loader_pvr;
 void register_pvr_types() {
 	resource_loader_pvr.instance();
 	ResourceLoader::add_resource_format_loader(resource_loader_pvr);
+
+	_register_pvrtc_compress_func();
 }
 
 void unregister_pvr_types() {
